Reworked structure in StaticAnalyszer. TODO: Move some shared_func from android over to shared_func and rework imports.
