+++ MobSF/urls.py
@@ -13,15 +13,15 @@ urlpatterns = [
     url(r'^NotFound/$', 'MobSF.views.NotFound', name='NotFound'),
     url(r'^ZIP_FORMAT/$', 'MobSF.views.ZIP_FORMAT', name='ZIP_FORMAT'),
     url(r'^MAC_ONLY/$', 'MobSF.views.MAC_ONLY', name='MAC_ONLY'),
-    url(r'^StaticAnalyzer/$', 'StaticAnalyzer.views.StaticAnalyzer', name='StaticAnalyzer'),
-    url(r'^StaticAnalyzer_iOS/$', 'StaticAnalyzer.views.StaticAnalyzer_iOS', name='StaticAnalyzer_iOS'),
-    url(r'^ViewSource/$', 'StaticAnalyzer.views.ViewSource', name='ViewSource'),
-    url(r'^PDF/$', 'StaticAnalyzer.views.PDF', name='PDF'),
-    url(r'^ViewFile/$', 'StaticAnalyzer.views.ViewFile', name='ViewFile'),
-    url(r'^Smali/$', 'StaticAnalyzer.views.Smali', name='Smali'),
-    url(r'^Java/$', 'StaticAnalyzer.views.Java', name='Java'),
+    url(r'^StaticAnalyzer/$', 'StaticAnalyzer.views.android.StaticAnalyzer', name='StaticAnalyzer'),
+    url(r'^StaticAnalyzer_iOS/$', 'StaticAnalyzer.views.ios.StaticAnalyzer_iOS', name='StaticAnalyzer_iOS'),
+    url(r'^ViewSource/$', 'StaticAnalyzer.views.android.ViewSource', name='ViewSource'),
+    url(r'^PDF/$', 'StaticAnalyzer.views.android.PDF', name='PDF'),
+    url(r'^ViewFile/$', 'StaticAnalyzer.views.android.ViewFile', name='ViewFile'),
+    url(r'^Smali/$', 'StaticAnalyzer.views.android.Smali', name='Smali'),
+    url(r'^Java/$', 'StaticAnalyzer.views.android.Java', name='Java'),
     url(r'^Find/$', 'StaticAnalyzer.views.Find', name='Find'),
-    url(r'^ManifestView/$', 'StaticAnalyzer.views.ManifestView', name='ManifestView'),
+    url(r'^ManifestView/$', 'StaticAnalyzer.views.android.ManifestView', name='ManifestView'),
     url(r'^DynamicAnalyzer/$', 'DynamicAnalyzer.views.DynamicAnalyzer', name='DynamicAnalyzer'),
     url(r'^GetEnv/$', 'DynamicAnalyzer.views.GetEnv', name='GetEnv'),
     url(r'^GetRes/$', 'DynamicAnalyzer.views.GetRes', name='GetRes'),
@@ -41,6 +41,3 @@ urlpatterns = [
     url(r'^NoAPI/$', 'APITester.views.NoAPI', name='NoAPI'),
     #url(r'^admin/', include(admin.site.urls)),
 ]
-
-
-
